Implement AcceptOrder and partially implement deleteLabel

+++ MHAuthorWebsite/MHAuthorWebsite.Core/Dto/EcontOrderDto.cs
@@ -2,7 +2,7 @@
 
 public class EcontOrderDto
 {
-    public long? Id { get; set; }
+    public int? Id { get; set; }
 
     public string? OrderNumber { get; set; }
